requirements.txt
nltk==3.5
pandas==1.1.5
numpy==1.17.3
pyserial==3.5
